requirements.txt
tensorflow-cpu
h5py==2.10.0
opencv-python-headless==4.2.0.32
numpy
gradio